requirements.txt
# Football Analytics Chatbot - Backend Requirements

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.6

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# ML/Models
scikit-learn>=1.3.0
joblib>=1.3.0

# LLM Integration (optional)
openai>=1.0.0
anthropic>=0.18.0

# Utilities
pydantic>=2.0.0
requests>=2.31.0

# Development
pytest>=7.4.0
httpx>=0.24.0
